I got the persistent server to work, I need to finish writing the test class for it
I added the jdbc driver to the classpath in the run_server file

I wrote a Database Setup class that creates all of the tables needed for the database to work correctly. The database is created at the path given in the first command line input
This class is run by the run_database_setup.sh

In the Panel files, I fixed some of the code to make the persistence work properly.

I added a Database class that connects to the database and has methods for saving messages, conversations, and users as well as allowing external classes to execute SQLite queries

In the Server class:
The constructor takes a Database object and uses it to construct the Controller

In the Server Controller class:
The constructor takes a Database object and stores it in a class variable, then loads all of the stored information from the database

all of the new___ methods have an additional check that fails if the object is not saved to the database. This allows the existing code for other failures to be used when saving to the database fails

I added a buildUuid function that takes an integer as a parameter and constructs a uuid with that as the id and the server id as the root. This is used for loading ids from the database

The load function loads the users, the conversations, and the messages, in that order.

For each user, the id, name, and creation time are loaded from the table and used to construct a User object that is added to the model

For each conversation, the id, title, owner id, and creation time are loaded from the table.
Then, the owner id is used to select the appropriate User object from the model.
Finally, the Conversation object is constructed and added to the model.

The messages are selected in order of creation time so that they are added to the conversations in the correct order
For each message, the id, creation time, author id, conversation id, and content
Then, the author id is used to select the appropriate User object from the model and the conversation id is used to select the appropriate Conversation object.
 Next, a new message is constructed and added to the model.
 Finally, the message is added to the appropriate linked list.

 After all of the loading is done, a list of all of the loaded data is logged.